requirements.txt
cerebras_cloud_sdk>=1.0.0
pydantic>=2.8.0
requests>=2.32.0
python-dotenv>=1.0.1